backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
pymupdf
pydantic
supabase==2.5.1
python-dotenv
httpx
openai 
requests 
scikit-learn==1.3.2  